add shared_example for Spot::WorksControllerBehavior
